requirements.txt
google-api-python-client
google-generativeai
psycopg2-binary
python-dotenv
pandas
numpy
streamlit
plotly
